Add patched `postgresql@12` formula to unblock development
There is a known issue w/ the `postgresql@12` formula being discussed
here: https://github.com/Homebrew/homebrew-core/issues/62930 with a
proposed fix here: https://github.com/Homebrew/homebrew-core/pull/62992

However until the above is merged, developers looking to install
`postgresql@12` are dead in the water. This is intended to only be a
temporary work-around and will be taken done once the fix is available
upstream.